backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
pillow>=10.3.0
ultralytics>=8.0.196
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
aiofiles>=23.2.1
boto3>=1.29.7
pytest>=7.4.3
httpx>=0.25.2
